+++ include/VkZero/Internal/frame_internal.hpp
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "VkZero/frame.hpp"
+
+namespace VkZero
+{
+struct FrameImpl_T {
+  FrameImpl_T(std::vector<RenderpassImpl_T *> renderpasses, WindowImpl_T* window);
+  void draw();
+
+  std::vector<VkSemaphore> imageAvailableSemaphores;
+  std::vector<VkSemaphore> renderFinishedSemaphores;
+  std::vector<VkFence> inFlightFences;
+  int currentFrame = 0;
+  std::vector<RenderpassImpl_T *> renderpasses;
+  std::vector<VkCommandBuffer> commandBuffers[MAX_FRAMES_IN_FLIGHT];
+  WindowImpl_T *window;
+  VkCommandPool commandPool;
+};
+}
